Autogenerate settings header
