+++ src/settings.cpp
@@ -32,7 +32,9 @@ with this program; if not, write to the Free Software Foundation, Inc.,
 #include "noise.h"
 #include <cctype>
 #include <algorithm>
+#include "settings_static.h"
 
+extern StaticSettingsManager builtin_settings_manager;
 static Settings main_settings;
 Settings *g_settings = &main_settings;
 std::string g_settings_path;
@@ -142,6 +144,7 @@ bool Settings::parseConfigLines(std::istream &is, const std::string &end)
 		case SPE_COMMENT:
 			break;
 		case SPE_KVPAIR:
+			builtin_settings_manager.update(name, value);
 			m_settings[name] = SettingsEntry(value);
 			break;
 		case SPE_END:
@@ -763,10 +766,16 @@ bool Settings::setEntry(const std::string &name, const void *data,
 
 		SettingsEntry &entry = set_default ? m_defaults[name] : m_settings[name];
 		old_group = entry.group;
-
-		entry.value    = set_group ? "" : *(const std::string *)data;
-		entry.group    = set_group ? *(Settings **)data : NULL;
 		entry.is_group = set_group;
+		if (set_group) {
+			entry.value.clear();
+			entry.group = *(Settings **)data;
+		} else {
+			entry.value = *(const std::string *)data;
+			entry.group = nullptr;
+			if (!set_default)
+				builtin_settings_manager.update(name, entry.value);
+		}
 	}
 
 	delete old_group;
